Explain how to import the lru_map package
And suggest adding `--save` to the `npm install`, because installing packages per-project rather than globally is a best practice, and there is already a simpler installation method listed for people who wouldn’t know that they need to run `npm init`.

It is necessary to explain how to import LRUMap because I at first tried to import it with `import LRUMap from 'lru_map';`. My build then gave only the cryptic error message “_lru_map2.default is not a constructor”. It took inspecting the LRUMap source code to figure out that I had to take the `LRUMap` property of what was exported instead of using the exported object directly.