fix type definitions
`entries()` was typed as returning an iterator, but it actually returns an iterable. (If it returned an iterator, you could call `map.entries().next()`, but that's a runtime error.)

Meanwhile, `keys()` and `values()` do return iterators, which makes their types correct but their behavior a bit inconsistent with `entries()`. I'm not sure if that inconsistency is an issue, but addressing it would require a breaking change, so I stuck to just fixing the types in this PR.